backend/requirements.txt
attrs==25.4.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.0.0
Flask-Cors==4.0.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==1.26.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.16.2
selenium==4.15.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
threadpoolctl==3.6.0
trio==0.31.0
trio-websocket==0.12.2
tzdata==2025.2
urllib3==2.5.0
webdriver-manager==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
